requirements.txt
addict
aenum
aiohttp
anyio
appdirs
astunparse
bitsandbytes
blendmodes
clean-fid
easydev
extcolors
facexlib
filetype
font-roboto
fonts
future
gdown
gfpgan
GitPython
httpcore
inflection
jsonmerge
kornia
lark
lmdb
lpips
omegaconf
open-clip-torch
opencv-contrib-python
piexif
Pillow
psutil
pyyaml
realesrgan
resize-right
rich
safetensors
scipy
tb_nightly
toml
torchdiffeq
torchsde
voluptuous
yapf
scikit-image
basicsr
compel
requests==2.31.0
tqdm==4.65.0
accelerate==0.18.0
opencv-python==4.7.0.72
diffusers==0.16.1
einops==0.4.1
gradio==3.32.0
numexpr==2.8.4
numpy==1.23.5
numba==0.57.0
pandas==1.5.3
protobuf==3.20.3
pytorch_lightning==1.9.4
transformers==4.26.1
timm==0.6.13
tomesd==0.1.2
urllib3==1.26.15
opencv-python==4.7.0.72
insightface==0.7.3
psutil==5.9.5
tk==0.1.0
pillow==9.5.0
codeformer-pip==0.0.4